+++ packages/pieces/apps/src/index.ts
@@ -6,8 +6,8 @@ import { binance } from '@activepieces/piece-binance';
 import { blackbaud } from '@activepieces/piece-blackbaud';
 import { calcom } from '@activepieces/piece-cal-com';
 import { calendly } from '@activepieces/piece-calendly';
+import { clickup } from '@activepieces/piece-clickup';
 import { slack } from '@activepieces/piece-slack';
-import { clickup } from './lib/clickup';
 import { discord } from './lib/discord';
 import { drip } from './lib/drip';
 import { dropbox } from './lib/dropbox';
