+++ packages/pieces/clickup/.babelrc
@@ -1,0 +1,3 @@
+{
+  "presets": [["@nrwl/js/babel", { "useBuiltIns": "usage" }]]
+}
