+++ packages/pieces/clickup/README.md
@@ -1,0 +1,7 @@
+# pieces-clickup
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running lint
+
+Run `nx lint pieces-clickup` to execute the lint via [ESLint](https://eslint.org/).
