+++ packages/pieces/clickup/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "@activepieces/piece-clickup",
+  "version": "0.0.1"
+}
