+++ packages/pieces/clickup/project.json
@@ -1,0 +1,27 @@
+{
+  "name": "pieces-clickup",
+  "$schema": "../../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "packages/pieces/clickup/src",
+  "projectType": "library",
+  "targets": {
+    "build": {
+      "executor": "@nrwl/js:tsc",
+      "outputs": ["{options.outputPath}"],
+      "options": {
+        "outputPath": "dist/packages/pieces/clickup",
+        "tsConfig": "packages/pieces/clickup/tsconfig.lib.json",
+        "packageJson": "packages/pieces/clickup/package.json",
+        "main": "packages/pieces/clickup/src/index.ts",
+        "assets": ["packages/pieces/clickup/*.md"]
+      }
+    },
+    "lint": {
+      "executor": "@nrwl/linter:eslint",
+      "outputs": ["{options.outputFile}"],
+      "options": {
+        "lintFilePatterns": ["packages/pieces/clickup/**/*.ts"]
+      }
+    }
+  },
+  "tags": []
+}
